+++ stac-api/src/lib.rs
@@ -102,3 +102,17 @@ pub type Result<T> = std::result::Result<T, Error>;
 /// used, so this specification provides a mechanism for clients to request that
 /// servers to explicitly include or exclude certain fields.
 pub type Item = serde_json::Map<String, serde_json::Value>;
+
+// From https://github.com/rust-lang/cargo/issues/383#issuecomment-720873790,
+// may they be forever blessed.
+#[cfg(doctest)]
+mod readme {
+    macro_rules! external_doc_test {
+        ($x:expr) => {
+            #[doc = $x]
+            extern "C" {}
+        };
+    }
+
+    external_doc_test!(include_str!("../README.md"));
+}